finally models

+++ models/hh_demographic.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, BigInteger, String
 
 from database import Base
 
@@ -13,7 +13,7 @@ class HHDemographic(Base):
     hh_comp_desc = Column(String(50), nullable=True)
     household_size_desc = Column(String(10), nullable=True)
     kid_category_desc = Column(String(30), nullable=True)
-    household_key = Column(Integer, primary_key=True, index=True)
+    household_key = Column(BigInteger, primary_key=True, index=True)
 
     def __repr__(self):
         return f"<HHDemographic(household_key={self.household_key}, age_desc='{self.age_desc}')>"
